+++ apps/coder/app.go
@@ -18,6 +18,7 @@ import (
 	"github.com/fluxplane/agentruntime/plugins/codingplugin"
 	"github.com/fluxplane/agentruntime/plugins/identityplugin"
 	"github.com/fluxplane/agentruntime/plugins/imageplugin"
+	"github.com/fluxplane/agentruntime/plugins/kubernetesplugin"
 	"github.com/fluxplane/agentruntime/plugins/skillplugin"
 	"github.com/fluxplane/agentruntime/plugins/taskplugin"
 	"github.com/fluxplane/agentruntime/plugins/workspaceplugin"
@@ -171,6 +172,7 @@ func localPlugins(hostSystem system.System) []pluginhost.Plugin {
 		taskplugin.New(),
 		skillplugin.New(),
 		imageplugin.New(hostSystem),
+		kubernetesplugin.New(hostSystem),
 	}
 }